test: add opt-in key parsing regression tests
